routes and role reform

+++ src/controllers/user.controller.ts
@@ -2,8 +2,9 @@ import { Request, Response } from "express";
 import User from "../models/user.model";
 import { redis } from "..";
 import sendResponse from "../middleware/sendResponse";
+import verifyToken from "../middleware/verifyToken";
 
-const getUsers = async (_req: Request, res: Response) => {
+const getUserByIds = async (_req: Request, res: Response) => {
   try {
     const cachedUsers = await redis.get("users");
     if (cachedUsers) {
@@ -15,17 +16,28 @@ const getUsers = async (_req: Request, res: Response) => {
         JSON.parse(cachedUsers),
       );
     }
-    const result = await User.find();
-    await redis.set("users", JSON.stringify(result));
-    return sendResponse(res, 200, true, "Get users successfully", result);
+    const user = await User.find();
+    await redis.set("users", JSON.stringify(user));
+    return sendResponse(res, 200, true, "Get users successfully", user);
   } catch (error) {
     return sendResponse(res, 500, false, error);
   }
 };
 
-const getUser = async (req: Request, res: Response) => {
+const getUserById = async (req: Request, res: Response) => {
   try {
     const { id } = req.params;
+    const authHeader = req.headers.authorization;
+    if (!authHeader || !authHeader.startsWith("Bearer ")) {
+      return sendResponse(res, 401, false, "Invalid token");
+    }
+    const decodedToken = verifyToken(authHeader);
+    if (!decodedToken) {
+      return sendResponse(res, 401, false, "You are unauthorized");
+    }
+    if (decodedToken.id !== id) {
+      return sendResponse(res, 401, false, "You are not owner");
+    }
     const cachedUser = await redis.get(`user:${id}`);
     if (cachedUser) {
       return sendResponse(
@@ -36,21 +48,32 @@ const getUser = async (req: Request, res: Response) => {
         JSON.parse(cachedUser),
       );
     }
-    const result = await User.findById(id);
-    if (!result) {
+    const user = await User.findById(id);
+    if (!user) {
       return sendResponse(res, 404, false, "User not found");
     }
-    await redis.set(`user:${id}`, JSON.stringify(result));
-    return sendResponse(res, 200, true, "Get user successfully", result);
+    await redis.set(`user:${id}`, JSON.stringify(user));
+    return sendResponse(res, 200, true, "Get user successfully", user);
   } catch (error) {
     return sendResponse(res, 500, false, error);
   }
 };
 
-const updateUser = async (req: Request, res: Response) => {
+const updateUserById = async (req: Request, res: Response) => {
   try {
     const { id } = req.params;
     const data = req.body;
+    const authHeader = req.headers.authorization;
+    if (!authHeader || !authHeader.startsWith("Bearer ")) {
+      return sendResponse(res, 401, false, "Invalid token");
+    }
+    const decodedToken = verifyToken(authHeader);
+    if (!decodedToken) {
+      return sendResponse(res, 401, false, "You are unauthorized");
+    }
+    if (decodedToken.id !== id) {
+      return sendResponse(res, 401, false, "You are not owner");
+    }
     const user = await User.findByIdAndUpdate(id, data, { new: true });
     if (!user) {
       return sendResponse(res, 404, false, "User not found");
@@ -63,9 +86,20 @@ const updateUser = async (req: Request, res: Response) => {
   }
 };
 
-const deleteUser = async (req: Request, res: Response) => {
+const deleteUserById = async (req: Request, res: Response) => {
   try {
     const { id } = req.params;
+    const authHeader = req.headers.authorization;
+    if (!authHeader || !authHeader.startsWith("Bearer ")) {
+      return sendResponse(res, 401, false, "Invalid token");
+    }
+    const decodedToken = verifyToken(authHeader);
+    if (!decodedToken) {
+      return sendResponse(res, 401, false, "You are unauthorized");
+    }
+    if (decodedToken.id !== id) {
+      return sendResponse(res, 401, false, "You are not owner");
+    }
     const user = await User.findByIdAndDelete(id);
     if (!user) {
       return sendResponse(res, 404, false, "User not found");
@@ -79,10 +113,10 @@ const deleteUser = async (req: Request, res: Response) => {
 };
 
 const userController = {
-  getUsers,
-  getUser,
-  updateUser,
-  deleteUser,
+  getUserByIds,
+  getUserById,
+  updateUserById,
+  deleteUserById,
 };
 
 export default userController;
